KVM: arm64 Check LRs instead of AP list for L1 irq injection
We used to check vgic_irq in the AP list to see if we need to inject an
irq to L1.  However, at this point, all the informations are flushed to
LRs.  So, let's look at LRs instead.

This fixes a problem that L0 doesn't inject IPIs to L1 even though it
has one for L1. This is because when populating LRs, the pending bit of
an edge sensitive IRQ (including IPI) is cleared in struct vgic_irq.

TODO: check the pending condition in the loop rather than after the loop.
list_for_each_entry(irq, &vgic_cpu->ap_list_head, ap_list)

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>