KVM: arm64: Sanitize the vcpu mode on eret emulation
Handle cases that the target exception level stored in the virtual
SPSR_EL2 is EL1; This could happen when the virtual EL2 trapped to
itself behind the host hypervisor's back or it returns to its VM.  For
the former, vcpu mode should stay at the virtual EL2 and for the latter,
vcpu mode is supposed to be EL1.

Currently, there's no good way to differentiate them clearly. Luckily,
there a temporal walk-around for the KVM guest hypervisor. The virtual
EL2 running KVM without VHE does not trap to itself, so we are safe. For
the virtual EL2 running KVM with VHE, we check the virtual TGE bit to
decide what exception level to return. Note that the virtual TGE bit is
not guranteed to be set in the virtual EL2 as the hypervisor
occasionally turns it off when executing tlbi instructions in the latest
kernel. Running v4.10.0 kernel as a guest hypervisor was fine.

Signed-off-by: Jintack Lim <jintack.lim@linaro.org>