KVM: arm/arm64: vgic: Update virtual distributor on EOI from a nested VM
When a nested OS deactivates a virtual interrupt, which is associated
with a hardware interrupt (i.e. HW bit set by the guest hypervisor),
then we reflect this to the virtual distributor for the guest
hypervisor.  This is equivalent to the interrupt state change made to
the physical distributor by hardware in the non-nesting case.

Without this patch, the interrupt state of L1 remains active and L0
can't inject further interrupts.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>