KVM: arm64: timer: Trap cntpct from VHE guest hypervisor
On VHE systems, cntvct and cntpct are the same in EL2. Therefore we no
longer allow the VHE guest hypervisor to access cntpct natively, but
emulate it to have the same value as cntvct.

Note that this may have a significant impact on performance.

TODO: We may need for non-VHE guest hypervisor as well so that the guest
hypervisor can see cntpct - cntvct = cntvoff. Use is_hyp_ctxt() instead
of vcpu_vhe_host() and test it for non-VHE hypervisor.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>