KVM: arm/arm64: Inject phys timer irqs on loading vcpus
We've been canceling and restarting a soft timer for the phys timer
emulation on vcpu_put and vcpu_load respectively. This would result in
not injecting phys timer irqs to the guest if the emulated phys timer is
expired while the soft timer is canceled. This behavior is always
observed when running a Linux guest modified to use the non-secure phys
timer, which prevents the guest from completing the booting process
waiting for a phys timer irq. This behavior is not catched by
kvm-unit-test, though.

To fix this problem, check if we need to inject a phys timer irq to the
guest on vcpu_load as well as on phys timer register modifications.

Fixes: d4151696e803 ("KVM: arm/arm64: Avoid phys timer emulation in vcpu entry/exit")
Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>