timer: Adjust virtual offset considering nested virtualization
For nested virtualization, we consider the cntvoff of the guest
hypervisor as well as the one of the host; the virtual CNTVOFF_EL2 is
set in the VM while the host already set a cntvoff_el2 for the VM, so we
to account both of the cntvoff_el2 set by the host and the virtual
cntvoff_el2.

Note that the VHE guest hypervisor needs a special attention.  When the
guest hypervisor is running (i.e. EL2 or EL0 of the host OS in the VM),
the virtual offset for the EL2 virt timer and cntvct is treated as 0
while the virtual offset for the EL1 timer is NOT. Since the access to
the EL2 virtual timer and cntvct are allowed without trap, we
set the phys cntvoff_el2 treating the virtual cntvoff_el2 as 0.

TODO: Check if we need to change set/get_reg() functions as well -
probably we do.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>