vgic/timer: Set GICH_VMCR for every vcpu entry
GICH_VMCR has VEM bit, which is an alias of GICV_CTLR.EOImode.
Currently we set GICH_VMCR on vcpu_load. However when we run nested
virtualization, EOImode can be changed while a vcpu is loaded. For
example, a guest hypervisor may set EOImode to its GICC, which is really
GICV while a nested OS may not set it. A particular problem I had was
that when deliverying a vtimer irq from L0 to L2 through L1, L1 keeps
getting vtimer irq repeatedly. This is because when L1 does EOI to GICV,
it deactivates active state, since EOImode was 0 when the vcpu was
loaded. This resulted in making vtimer irq pending again, so L1 (and L2)
couldn't proceed.

So, let's update VMCR on every vcpu entry. We may optimized it to be
done when there's a switch between the guest hypervisor and a nested VM.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>