memory: Close hw_vttbr race
Whenever we get a new vmid, we update hw_vttbr to have a new vmid. This
is *only* necessary when we don't use nested virtualization. When we do
use nested virtualization, we always set hw_vttbr before entering the
guest, so this is not necessary. Updating hw_vttbr when using nested
virtualization is not correct either. For example, when we get a new
vmid for a nested VM, updating hw_vttbr may overwrite existing hw_vttbr
value which might be used to enter L1 VM.

This was the culprit of getting memory-related crash for L0, L1 and L2
occasionally.

Fixes: e077a15a9b83 ("KVM: arm/arm64: Support mmu for the virtual EL2 execution")

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>